+++ lib/tmuxinator/project.rb
@@ -47,7 +47,8 @@ module Tmuxinator
         @settings = parse_settings(args)
         @args = args
 
-        content = Erubis::Eruby.new(raw_content).result(binding)
+        erubi_content = Erubi::Engine.new(raw_content).src
+        content = binding.instance_eval(erubi_content)
         YAML.safe_load(content, aliases: true)
       rescue SyntaxError, StandardError => error
         raise "Failed to parse config file: #{error.message}"
@@ -103,7 +104,7 @@ module Tmuxinator
 
     def self.render_template(template, bndg)
       content = File.read(template)
-      Erubis::Eruby.new(content).result(bndg)
+      bndg.eval(Erubi::Engine.new(content).src)
     end
 
     def windows