+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Replace erubis with erubi
+
 ## 3.0.0
 ### Misc
 - Deprecate Ruby 2.5; bump min Ruby version in gemspec; bump Ruby versions in Travis test matrix